+++ app/Http/Controllers/V1/RealEstate/UnitController.php
@@ -11,6 +11,10 @@ use App\Services\RealEstate\UnitService;
 use App\Traits\ResponseTrait;
 use Illuminate\Http\Request;
 
+/**
+ * @method void authorize(mixed $ability, mixed|array $arguments = [])
+ */
+
 class UnitController extends Controller
 {
     use ResponseTrait;
@@ -20,12 +24,14 @@ class UnitController extends Controller
 
     public function index(int $building_id)
     {
+        $this->authorize('view');
         $units = $this->unitService->index($building_id, []);
         return $this->successCollection($units, UnitResource::class);
     }
 
     public function store(CreateUnitRequest $request)
     {
+        $this->authorize('create');
         $unitDTO = CreateUnitDTO::fromRequest($request->validated());
         $unit = $this->unitService->store($unitDTO);
         return $this->useResource($unit, UnitResource::class, __('messages.common.stored'), 201);
@@ -33,6 +39,7 @@ class UnitController extends Controller
 
     public function show(int $id)
     {
+        $this->authorize('view');
         $unit = $this->unitService->show($id);
         return $this->useResource($unit, UnitResource::class);
     }
@@ -45,13 +52,15 @@ class UnitController extends Controller
 
     public function update(int $id, Request $request)
     {
+        $this->authorize('update');
         $unitDTO = UpdateUnitDTO::fromRequest($request->all());
         $unit = $this->unitService->update($id, $unitDTO);
         return $this->useResource($unit, UnitResource::class, __('messages.common.updated'));
     }
 
-    public function destroy($id)
+    public function destroy(int $id)
     {
+        $this->authorize('destroy');
         $this->unitService->destroy($id);
         return $this->successResponse([], __('messages.common.deleted'));
     }
